requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
serpapi
langchain-core
langchain-groq
python-dotenv
gspread-pandas
google-search-results